Hacks/Extensions to PFsLib - volume name, faster free cluster count
Started to migrate some of the sketch stuff out of the sketch and into
PFsVolume stuff.

Like read the volume label.

Also overwwrite of the freeClusterCount.

This took some hacks to need to cast to the BlockDevice class for the USB, such that I know we are using that and as such can call the red Sectors with callbacks.