WIP - decouple MSC from USBHost
Unclear at this point if this code will be ever integrated directly into the USBHost library , at least in the current form.

As other things are getting fixed within the usbhost code and it is difficult to try to maintain two working versions, I will probably close out the integration branch

This is first pass at migrating the code back into this library.

As a complication, I renamamed the msController objects within this library to msDevice as to decouple any requirements on the usbhost code versions.  i.e. sort of orphaned the msController class